pybfm/obj: Move mesh buffer functions here
