+++ pybfm/bfm/instance.py
@@ -24,8 +24,6 @@ class CInstance:
 	# visualisation stuff
 
 	def gen_buffers(self):
-		mesh = self.obj.mesh
-
 		# create VAO
 
 		self.vao = gl.GLuint(0)
@@ -38,11 +36,11 @@ class CInstance:
 		gl.glGenBuffers(1, ctypes.byref(self.vbo))
 		gl.glBindBuffer(gl.GL_ARRAY_BUFFER, self.vbo)
 
-		coords_t = gl.GLfloat * len(mesh.gl_coords)
+		coords_t = gl.GLfloat * len(self.obj.gl_coords)
 
 		gl.glBufferData(gl.GL_ARRAY_BUFFER,
 			ctypes.sizeof(coords_t),
-			(coords_t) (*mesh.gl_coords),
+			(coords_t) (*self.obj.gl_coords),
 			gl.GL_STATIC_DRAW)
 
 		gl.glVertexAttribPointer(0, 3, gl.GL_FLOAT, gl.GL_FALSE, 0, 0)
@@ -64,11 +62,11 @@ class CInstance:
 		gl.glGenBuffers(1, self.ibo)
 		gl.glBindBuffer(gl.GL_ELEMENT_ARRAY_BUFFER, self.ibo)
 
-		indices_t = gl.GLuint * len(mesh.indices)
+		indices_t = gl.GLuint * len(self.obj.indices)
 
 		gl.glBufferData(gl.GL_ELEMENT_ARRAY_BUFFER,
 			ctypes.sizeof(indices_t),
-			(indices_t) (*mesh.indices),
+			(indices_t) (*self.obj.indices),
 			gl.GL_STATIC_DRAW)
 
 		# create lines IBO
@@ -77,11 +75,11 @@ class CInstance:
 		gl.glGenBuffers(1, self.lines_ibo)
 		gl.glBindBuffer(gl.GL_ELEMENT_ARRAY_BUFFER, self.lines_ibo)
 
-		line_indices_t = gl.GLuint * len(mesh.line_indices)
+		line_indices_t = gl.GLuint * len(self.obj.line_indices)
 
 		gl.glBufferData(gl.GL_ELEMENT_ARRAY_BUFFER,
 			ctypes.sizeof(line_indices_t),
-			(line_indices_t) (*mesh.line_indices),
+			(line_indices_t) (*self.obj.line_indices),
 			gl.GL_STATIC_DRAW)
 
 	@functools.cached_property
@@ -100,7 +98,7 @@ class CInstance:
 		mesh = self.obj.mesh
 		max_effect = 0
 
-		for i in range(mesh.c_mesh.n_nodes):
+		for i in range(self.obj.mesh.c_mesh.n_nodes):
 			x = self.c_instance.effects[i * 2 + 0]
 			y = self.c_instance.effects[i * 2 + 1]
 
@@ -122,17 +120,15 @@ class CInstance:
 
 	def draw(self, shader: Shader, lines = False):
 		shader.max_effect(self.max_effect)
-
-		mesh = self.obj.mesh
 		gl.glBindVertexArray(self.vao)
 
 		if lines:
 			gl.glBindBuffer(gl.GL_ELEMENT_ARRAY_BUFFER, self.lines_ibo)
-			gl.glDrawElements(gl.GL_LINES, len(mesh.line_indices), gl.GL_UNSIGNED_INT, None)
+			gl.glDrawElements(gl.GL_LINES, len(self.obj.line_indices), gl.GL_UNSIGNED_INT, None)
 
 		else:
 			gl.glBindBuffer(gl.GL_ELEMENT_ARRAY_BUFFER, self.ibo)
-			gl.glDrawElements(gl.GL_TRIANGLES, len(mesh.indices), gl.GL_UNSIGNED_INT, None)
+			gl.glDrawElements(gl.GL_TRIANGLES, len(self.obj.indices), gl.GL_UNSIGNED_INT, None)
 
 class Instance(CInstance):
 	def __init__(self, obj: Obj):
